+++ tests/QueryTest.php
@@ -1136,5 +1136,89 @@ class QueryTest extends \PHPUnit\Framework\TestCase
 
         $db->flush(true);
     }
+    public function testDeleteWhereMatchItemsWithCustomFilter()
+    {
+        $db = new \Filebase\Database([
+            'dir' => __DIR__.'/databases/deletefilter',
+        ]);
+        $db->flush(true);
+
+        $a=0;
+        while($a < 15)
+        {
+            $user=$db->get($a."username");
+            $user->name  = $a.'John';
+            $user->email = 'john@example.com';
+            $user->tags  = ['php','developer','html5'];
+
+    		$user->save();
+            $a++;
+        }
+        $actual=$db->query()->results();
+        $this->assertCount(15,$actual);
+        $r=$db->query()->where('name','LIKE','john')->resultDocuments();
+        $this->assertInstanceOf(Document::class, $r[0]);
+
+        $db->query()->where('name','LIKE','john')->delete(function($item){
+
+            return $item->name=='0John';
+
+        });
+        $actual=$db->query()->where('name','LIKE','john')->resultDocuments();
+        $this->assertCount(14,$actual);
+        $db->flush(true);
+
+    }
+
+
+    public function testSortByTimes()
+    {
+        $db = new \Filebase\Database([
+            'dir' => __DIR__.'/databases/_testsort'
+        ]);
+
+        $db->flush(true);
+
+        // Create some docs with time in between to get different timestamps
+        $doc = $db->get('record1')->set(['name'=>'a'])->save();
+        sleep(1);
+        $doc = $db->get('record2')->set(['name'=>'b'])->save();
+        sleep(1);
+        $doc = $db->get('record3')->set(['name'=>'c'])->save();
+
+        $documents = $db->query()->orderBy('__created_at', 'DESC')->results();
+        $expected = [
+            ['name' => 'c'],
+            ['name' => 'b'],
+            ['name' => 'a'],
+        ];
+        $this->assertEquals($expected, $documents);
+
+        $documents = $db->query()->orderBy('__created_at', 'ASC')->results();
+        $expected = [
+            ['name' => 'a'],
+            ['name' => 'b'],
+            ['name' => 'c'],
+        ];
+        $this->assertEquals($expected, $documents);
+
+        $documents = $db->query()->orderBy('__updated_at', 'DESC')->results();
+        $expected = [
+            ['name' => 'c'],
+            ['name' => 'b'],
+            ['name' => 'a'],
+        ];
+        $this->assertEquals($expected, $documents);
+
+        $documents = $db->query()->orderBy('__updated_at', 'ASC')->results();
+        $expected = [
+            ['name' => 'a'],
+            ['name' => 'b'],
+            ['name' => 'c'],
+        ];
+        $this->assertEquals($expected, $documents);
+
+        $db->flush(true);
+    }
 
 }